refactor: extract calendar helpers to shared lib
