+++ packages/shared/src/lib/calendar.ts
@@ -1,0 +1,90 @@
+export type AddToCalendarProvider = 'google' | 'outlook' | 'ics';
+
+export interface CalendarEvent {
+  title: string;
+  description: string;
+  location: string;
+  start: Date;
+  end: Date;
+  id: string;
+}
+
+const formatCalendarDate = (date: Date): string =>
+  `${date.toISOString().replace(/[-:]/g, '').split('.')[0]}Z`;
+
+const escapeIcsText = (value: string): string =>
+  value
+    .replace(/\\/g, '\\\\')
+    .replace(/\n/g, '\\n')
+    .replace(/,/g, '\\,')
+    .replace(/;/g, '\\;');
+
+export const buildGoogleCalendarUrl = (event: CalendarEvent): string => {
+  const params = new URLSearchParams({
+    action: 'TEMPLATE',
+    text: event.title,
+    dates: `${formatCalendarDate(event.start)}/${formatCalendarDate(
+      event.end,
+    )}`,
+    details: event.description,
+    location: event.location,
+  });
+  return `https://calendar.google.com/calendar/render?${params.toString()}`;
+};
+
+export const buildOutlookCalendarUrl = (event: CalendarEvent): string => {
+  const params = new URLSearchParams({
+    path: '/calendar/action/compose',
+    rru: 'addevent',
+    startdt: event.start.toISOString(),
+    enddt: event.end.toISOString(),
+    subject: event.title,
+    body: event.description,
+    location: event.location,
+  });
+  return `https://outlook.live.com/calendar/0/deeplink/compose?${params.toString()}`;
+};
+
+export const buildIcsContent = (event: CalendarEvent): string =>
+  [
+    'BEGIN:VCALENDAR',
+    'VERSION:2.0',
+    'PRODID:-//daily.dev//Calendar//EN',
+    'CALSCALE:GREGORIAN',
+    'METHOD:PUBLISH',
+    'BEGIN:VEVENT',
+    `UID:${event.id}@daily.dev`,
+    `DTSTAMP:${formatCalendarDate(new Date())}`,
+    `DTSTART:${formatCalendarDate(event.start)}`,
+    `DTEND:${formatCalendarDate(event.end)}`,
+    `SUMMARY:${escapeIcsText(event.title)}`,
+    `DESCRIPTION:${escapeIcsText(event.description)}`,
+    `LOCATION:${escapeIcsText(event.location)}`,
+    `URL:${event.location}`,
+    'END:VEVENT',
+    'END:VCALENDAR',
+  ].join('\r\n');
+
+interface DownloadIcsOptions {
+  filename?: string;
+}
+
+export const downloadIcs = (
+  event: CalendarEvent,
+  { filename }: DownloadIcsOptions = {},
+): void => {
+  if (typeof window === 'undefined') {
+    return;
+  }
+  const blob = new Blob([buildIcsContent(event)], {
+    type: 'text/calendar;charset=utf-8',
+  });
+  const url = URL.createObjectURL(blob);
+  const anchor = document.createElement('a');
+  anchor.href = url;
+  anchor.download = filename ?? `${event.id}.ics`;
+  document.body.appendChild(anchor);
+  anchor.click();
+  document.body.removeChild(anchor);
+  URL.revokeObjectURL(url);
+};
